backend/requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
django-cors-headers==3.14.0
flake8~=6.0.0
coverage
django-nose
djangorestframework-simplejwt==5.2.2
Faker==17.6.0
factory-boy==3.2.1